requirements.txt
streamlit>=1.28.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
bcrypt>=4.0.0
razorpay>=1.3.0